+++ projects/igniteui-angular/src/lib/grid/grid-header.component.ts
@@ -17,9 +17,8 @@ import { DataType } from '../data-operations/data-util';
 import { SortingDirection } from '../data-operations/sorting-expression.interface';
 import { RestrictDrag } from '../directives/dragdrop/dragdrop.directive';
 import { IgxGridAPIService } from './api.service';
-import { IgxGridCellComponent } from './cell.component';
 import { IgxColumnComponent } from './column.component';
-import { autoWire, IGridBus, IgxColumnMovingService } from './grid.common';
+import { IgxColumnMovingService } from './grid.common';
 
 @Component({
     changeDetection: ChangeDetectionStrategy.OnPush,
@@ -27,7 +26,7 @@ import { autoWire, IGridBus, IgxColumnMovingService } from './grid.common';
     selector: 'igx-grid-header',
     templateUrl: './grid-header.component.html'
 })
-export class IgxGridHeaderComponent implements IGridBus, OnInit, DoCheck, AfterViewInit {
+export class IgxGridHeaderComponent implements OnInit, DoCheck, AfterViewInit {
 
     @Input()
     public column: IgxColumnComponent;
@@ -37,6 +36,9 @@ export class IgxGridHeaderComponent implements IGridBus, OnInit, DoCheck, AfterV
 
     @HostBinding('class')
     get styleClasses() {
+        if (this.column.columnGroup) {
+            return `${this.column.headerClasses}`;
+        }
         return `igx-grid__th ${this.column.headerClasses}`;
     }
 
@@ -45,6 +47,14 @@ export class IgxGridHeaderComponent implements IGridBus, OnInit, DoCheck, AfterV
     @HostBinding('class.igx-grid__th--fw')
     get width() {
         return this.column.width;
+    }
+
+    @HostBinding('style.height.px')
+    get height() {
+        if (this.grid.hasColumnGroups) {
+            return (this.grid.maxLevelHeaderDepth + 1 - this.column.level) * this.grid.defaultRowHeight;
+        }
+        return null;
     }
 
     @HostBinding('class.asc')
@@ -131,15 +141,17 @@ export class IgxGridHeaderComponent implements IGridBus, OnInit, DoCheck, AfterV
     }
 
     ngAfterViewInit() {
-        this.zone.runOutsideAngular(() => {
-            this.resizeArea.nativeElement.addEventListener('mouseover', this.onResizeAreaMouseOver.bind(this));
-            this.resizeArea.nativeElement.addEventListener('mousedown', this.onResizeAreaMouseDown.bind(this));
-        });
+        if (!this.column.columnGroup) {
+            this.zone.runOutsideAngular(() => {
+                this.resizeArea.nativeElement.addEventListener('mouseover', this.onResizeAreaMouseOver.bind(this));
+                this.resizeArea.nativeElement.addEventListener('mousedown', this.onResizeAreaMouseDown.bind(this));
+            });
+        }
     }
 
     @HostListener('click', ['$event'])
-    @autoWire(true)
     public onClick(event) {
+
         if (!this.column.grid.isColumnResizing) {
             event.stopPropagation();
             if (this.column.sortable) {
@@ -177,7 +189,8 @@ export class IgxGridHeaderComponent implements IGridBus, OnInit, DoCheck, AfterV
         const actualWidth = this.elementRef.nativeElement.getBoundingClientRect().width;
 
         if (this.column.pinned) {
-            const pinnedMaxWidth = this._pinnedMaxWidth = this.grid.calcPinnedContainerMaxWidth - this.grid.pinnedWidth + actualWidth;
+            const pinnedMaxWidth = this._pinnedMaxWidth =
+                this.grid.calcPinnedContainerMaxWidth - this.grid.getPinnedWidth(true) + actualWidth;
 
             if (this.column.maxWidth && parseFloat(this.column.maxWidth) < pinnedMaxWidth) {
                 this._pinnedMaxWidth = this.column.maxWidth;
@@ -288,7 +301,7 @@ export class IgxGridHeaderComponent implements IGridBus, OnInit, DoCheck, AfterV
             const size = Math.ceil(largestCell + largestCellPadding) + 'px';
 
             if (this.column.pinned) {
-                const newPinnedWidth = this.grid.pinnedWidth - currentColWidth + parseFloat(size);
+                const newPinnedWidth = this.grid.getPinnedWidth(true) - currentColWidth + parseFloat(size);
 
                 if (newPinnedWidth <= this.grid.calcPinnedContainerMaxWidth) {
                     this.column.width = size;