+++ projects/igniteui-angular/src/lib/grid/grid-toolbar.component.ts
@@ -7,37 +7,45 @@ import {
     ViewChild
 } from '@angular/core';
 
-import { IgxToggleDirective } from '../directives/toggle/toggle.directive';
+import { DisplayDensity } from '../core/utils';
 import { CsvFileTypes,
          IgxBaseExporter,
          IgxCsvExporterOptions,
          IgxCsvExporterService,
          IgxExcelExporterOptions,
-         IgxExcelExporterService } from '../services/index';
+         IgxExcelExporterService,
+         CloseScrollStrategy} from '../services/index';
 import { IgxGridAPIService } from './api.service';
-import { IGridBus } from './grid.common';
 import { IgxGridComponent } from './grid.component';
 import { IgxDropDownComponent } from '../drop-down/drop-down.component';
 import { IgxColumnHidingComponent } from './column-hiding.component';
+import { OverlaySettings, PositionSettings, HorizontalAlignment, VerticalAlignment } from '../services/overlay/utilities';
+import {  ConnectedPositioningStrategy } from '../services/overlay/position';
 
 @Component({
     selector: 'igx-grid-toolbar',
     templateUrl: './grid-toolbar.component.html'
 })
-export class IgxGridToolbarComponent implements IGridBus {
+export class IgxGridToolbarComponent {
 
     @HostBinding('class.igx-grid-toolbar')
     @Input()
     public gridID: string;
-
-    @ViewChild(IgxToggleDirective, { read: IgxToggleDirective })
-    protected toggleDirective: IgxToggleDirective;
 
     @ViewChild('columnHidingDropdown', { read: IgxDropDownComponent })
     public columnHidingDropdown: IgxDropDownComponent;
 
     @ViewChild(IgxColumnHidingComponent)
     public columnHidingUI: IgxColumnHidingComponent;
+
+    @ViewChild('columnHidingButton')
+    public columnHidingButton;
+
+    @ViewChild('exportDropdown', { read: IgxDropDownComponent })
+    public exportDropdown: IgxDropDownComponent;
+
+    @ViewChild('btnExport')
+    public exportButton;
 
     public get grid(): IgxGridComponent {
         return this.gridAPI.get(this.gridID);
@@ -55,11 +63,60 @@ export class IgxGridToolbarComponent implements IGridBus {
         return (this.grid != null && this.grid.exportCsv);
     }
 
+    private _displayDensity: DisplayDensity | string;
+
+    @Input()
+    public get displayDensity(): DisplayDensity | string {
+        return this._displayDensity;
+    }
+
+    public set displayDensity(val: DisplayDensity | string) {
+        switch (val) {
+            case 'compact':
+                this._displayDensity = DisplayDensity.compact;
+                break;
+            case 'cosy':
+                this._displayDensity = DisplayDensity.cosy;
+                break;
+            case 'comfortable':
+            default:
+                this._displayDensity = DisplayDensity.comfortable;
+        }
+    }
+
+    @HostBinding('attr.class')
+    get hostClass(): string {
+        switch (this._displayDensity) {
+            case DisplayDensity.compact:
+                return 'igx-grid-toolbar--compact';
+            case DisplayDensity.cosy:
+                return 'igx-grid-toolbar--cosy';
+            case DisplayDensity.comfortable:
+            default:
+                return 'igx-grid-toolbar';
+        }
+    }
+
     constructor(public gridAPI: IgxGridAPIService,
                 public cdr: ChangeDetectorRef,
                 @Optional() public excelExporter: IgxExcelExporterService,
                 @Optional() public csvExporter: IgxCsvExporterService) {
     }
+
+    private _positionSettings: PositionSettings = {
+        horizontalDirection: HorizontalAlignment.Left,
+        horizontalStartPoint: HorizontalAlignment.Right,
+        verticalDirection: VerticalAlignment.Bottom,
+        verticalStartPoint: VerticalAlignment.Bottom
+    };
+
+    private _overlaySettings: OverlaySettings = {
+        positionStrategy: new ConnectedPositioningStrategy(this._positionSettings),
+        scrollStrategy: new CloseScrollStrategy(),
+        modal: false,
+        closeOnOutsideClick: true
+    };
+
 
     public getTitle(): string {
         return this.grid != null ? this.grid.toolbarTitle : '';
@@ -78,7 +135,9 @@ export class IgxGridToolbarComponent implements IGridBus {
     }
 
     public exportClicked() {
-        this.toggleDirective.collapsed = !this.toggleDirective.collapsed;
+
+        this._overlaySettings.positionStrategy.settings.target = this.exportButton.nativeElement;
+        this.exportDropdown.toggle(this._overlaySettings);
     }
 
     public exportToExcelClicked() {
@@ -107,6 +166,7 @@ export class IgxGridToolbarComponent implements IGridBus {
     }
 
     public toggleColumnHidingUI() {
-        this.columnHidingDropdown.toggle();
+        this._overlaySettings.positionStrategy.settings.target = this.columnHidingButton.nativeElement;
+        this.columnHidingDropdown.toggle(this._overlaySettings);
     }
 }