+++ src/app/routing.ts
@@ -42,9 +42,11 @@ import { GridSelectionComponent } from './grid-selection/grid-selection.sample';
 import { GridToolbarSampleComponent } from './grid-toolbar/grid-toolbar.sample';
 import { GridVirtualizationSampleComponent } from './grid-remote-virtualization/grid-remote-virtualization.sample';
 import { ButtonGroupSampleComponent } from './buttonGroup/buttonGroup.sample';
+import { GridColumnGroupsSampleComponent } from './grid-column-groups/grid-column-groups.sample';
 import { DropDownSampleComponent } from './drop-down/drop-down.sample';
 import { ComboSampleComponent } from './combo/combo.sample';
 import { OverlaySampleComponent } from './overlay/overlay.sample';
+import { RadioSampleComponent } from './radio/radio.sample';
 
 const appRoutes = [
     {
@@ -145,6 +147,10 @@ const appRoutes = [
         component: ProgressbarSampleComponent
     },
     {
+        path: 'radio',
+        component: RadioSampleComponent
+    },
+    {
         path: 'ripple',
         component: RippleSampleComponent
     },
@@ -232,6 +238,10 @@ const appRoutes = [
     {
         path: 'buttonGroup',
         component: ButtonGroupSampleComponent
+    },
+    {
+        path: 'gridColumnGroups',
+        component: GridColumnGroupsSampleComponent
     },
     {
         path: 'gridGroupBy',